+++ src/person/serializers.py
@@ -41,7 +41,6 @@ class UserSerializer(serializers.ModelSerializer):
     graduation_year = serializers.CharField(source="person.graduation_year")
     pronouns = serializers.CharField(source="person.pronouns")
     goals = SerializerMethodField("get_goals")
-    talking_points = SerializerMethodField("get_talking_points")
     availability = SerializerMethodField("get_availability")
     locations = LocationSerializer(source="person.locations", many=True)
     interests = InterestSerializer(source="person.interests", many=True)
@@ -56,12 +55,6 @@ class UserSerializer(serializers.ModelSerializer):
             return []
         goals = json.loads(user.person.goals)
         return goals
-
-    def get_talking_points(self, user):
-        if user.person.talking_points is None:
-            return []
-        talking_points = json.loads(user.person.talking_points)
-        return talking_points
 
     def get_availability(self, user):
         if user.person.availability is None:
@@ -112,7 +105,6 @@ class UserSerializer(serializers.ModelSerializer):
             "graduation_year",
             "pronouns",
             "goals",
-            "talking_points",
             "availability",
             "locations",
             "interests",
@@ -121,34 +113,6 @@ class UserSerializer(serializers.ModelSerializer):
             "has_onboarded",
             "pending_feedback",
             "current_match",
-        )
-        read_only_fields = fields
-
-
-class SimpleUserSerializer(serializers.ModelSerializer):
-    """Serializer for all users view."""
-
-    net_id = serializers.CharField(source="person.net_id")
-    profile_pic_url = serializers.CharField(source="person.profile_pic_url")
-    majors = MajorSerializer(source="person.majors", many=True)
-    hometown = serializers.CharField(source="person.hometown")
-    graduation_year = serializers.CharField(source="person.graduation_year")
-    interests = InterestSerializer(source="person.interests", many=True)
-    groups = GroupSerializer(source="person.groups", many=True)
-
-    class Meta:
-        model = User
-        fields = (
-            "id",
-            "net_id",
-            "first_name",
-            "last_name",
-            "profile_pic_url",
-            "majors",
-            "hometown",
-            "graduation_year",
-            "interests",
-            "groups",
         )
         read_only_fields = fields
