+++ src/person/simple_serializers.py
@@ -1,17 +1,20 @@
 from django.contrib.auth.models import User
+from group.serializers import GroupSerializer
+from interest.serializers import InterestSerializer
+from major.serializers import MajorSerializer
 from rest_framework import serializers
 
 
-class MatchUserSerializer(serializers.ModelSerializer):
-    """User serializer that excludes match history."""
+class SimpleUserSerializer(serializers.ModelSerializer):
+    """Serializer for all users view."""
 
     net_id = serializers.CharField(source="person.net_id")
-    hometown = serializers.CharField(source="person.hometown")
     profile_pic_url = serializers.CharField(source="person.profile_pic_url")
-    facebook_url = serializers.CharField(source="person.facebook_url")
-    instagram_username = serializers.CharField(source="person.instagram_username")
+    majors = MajorSerializer(source="person.majors", many=True)
+    hometown = serializers.CharField(source="person.hometown")
     graduation_year = serializers.CharField(source="person.graduation_year")
-    pronouns = serializers.CharField(source="person.pronouns")
+    interests = InterestSerializer(source="person.interests", many=True)
+    groups = GroupSerializer(source="person.groups", many=True)
 
     class Meta:
         model = User
@@ -20,11 +23,11 @@ class MatchUserSerializer(serializers.ModelSerializer):
             "net_id",
             "first_name",
             "last_name",
-            "hometown",
             "profile_pic_url",
-            "facebook_url",
-            "instagram_username",
+            "majors",
+            "hometown",
             "graduation_year",
-            "pronouns",
+            "interests",
+            "groups",
         )
         read_only_fields = fields
